+++ tests/Graph/Trees/ITreeTest.php
@@ -17,6 +17,8 @@ use PHPUnit\Framework\TestCase;
  */
 class ITreeTest extends TestCase
 {
+    protected const DATASET_SIZE = 100;
+
     protected const RANDOM_SEED = 0;
 
     /**
@@ -59,7 +61,7 @@ class ITreeTest extends TestCase
      */
     public function growSearch() : void
     {
-        $this->tree->grow($this->generator->generate(50));
+        $this->tree->grow($this->generator->generate(self::DATASET_SIZE));
 
         $this->assertGreaterThan(5, $this->tree->height());
 
@@ -69,6 +71,22 @@ class ITreeTest extends TestCase
 
         $this->assertInstanceOf(Depth::class, $node);
         $this->assertInstanceOf(BinaryNode::class, $node);
+    }
+
+    /**
+     * @test
+     */
+    public function growWithSameSamples() : void
+    {
+        $generator = new Agglomerate([
+            'east' => new Blob([5, -2, 10], 0.0),
+        ]);
+
+        $dataset = $generator->generate(self::DATASET_SIZE);
+
+        $this->tree->grow($dataset);
+
+        $this->assertEquals(2, $this->tree->height());
     }
 
     protected function assertPreConditions() : void